[Essential]Move Adapters to its own files
